More refactoring of lambda lowering for coroutines
Summary: We're getting very close to getting this right. We now have code that correctly generates the parameters and properties of the closure class. But before it can be turned on, we'll need to do some refactoring to get the generation of the constructor call correct. (Since we are changing the properties.)

Reviewed By: michaeltingley

Differential Revision: D5513481

fbshipit-source-id: 67b278a18c5bdc0502d9c63ed6fae7f4c2f8e1fc