we don't support completionItem/resolve\!
Summary:
Hack LSP server doesn't currently support completionItem/resolve. (That will become useful in the future when we start to provide doc-comments... but we don't provide them yet).

It currently claims that it does support completionItem/resolve. That results in resolve-aware clients like VSCode sending resolve requests, and getting back an error that the method isn't implemented.

This diff removes that claim.

Reviewed By: bolinfest

Differential Revision: D5477646

fbshipit-source-id: ae03e0e08caf9d0fb16a973a03af306fb12a172c