Optimize comparisons where both inputs are narrow
Summary:
If both inputs to a compare are known to have zeros above 8, 16 or 32
bits, we can replace the compare with a suitably sized one - as long
as we're careful about signs. If the result of the compare was used in
a "signed" sense, the zero extension of its inputs would actually make
it an unsigned compare - so when we reduce it to a narrow compare we
also have to change the uses of its result to be unsigned.

There was a lot of missing support for 16 bit types (eg cmpwim but no
cmpw or cmpwi), so I added a number of new instructions.

Reviewed By: mxw

Differential Revision: D5499172

fbshipit-source-id: 967e1f800d0926dfc06e39ec81791b6df97853da