improvements to fuzzer genetic programming
Summary:
This update adds a single-threaded implementation of coverage-based generational pruning to the fuzzer. Using llvm's profdata/cov system, it will build HHVM in coverage mode using Buck, and then after each run of a mutation, will perform an additional check to see if new coverage was added by the mutation. If the mutation only exercises lines already covered by previous mutations, it is discarded. A multi-threaded implementation is coming, but is currently non-functional for reasons being investigated.

In order to have Buck gather the coverage data, pertinent files to the fuzzer were separated into separate TARGETS files. Because Buck only allows a user to specify a directory in which to look for TARGETS files for which it will collect coverage data, if all of HHVM runtime was built in a single TARGETS file a great deal of unnecessary data would be collected, thus slowing down the fuzzer's speed.

A few small changes were also made to both the hack hhas parser and the fuzzer to allow the fuzzer to round-trip with itself. It will now almost never generate mutations that cannot be parsed as input to the fuzzer as part of a new generation.

Reviewed By: paulbiss

Differential Revision: D5483863

fbshipit-source-id: bb2f76534fb7a5b55b536ca8d4fb2f1d5e7a0d10