Creating a new ExtraTokenError trivia kind
Summary:
This diff is one of a stack that implements a new error recovery scheme for extra tokens faced by the full fidelity parser. The FFP needs to systematically handle these extra tokens somehow, in order to maintain the invariant that an in-order-traversal of the parsed SyntaxTree give back exactly the text of the source file. More context is quoted below, but--all this particular diff does is create the new ExtraTokenError trivia kind, and ensure compatibility with existing code (like by regenerating full_fidelity_schema.ml, and adding some cases to the match statements in hackformat.)

jakebailey like I mentioned before, I don't think any further work from you is necessary here, since we're just treating these error trivia like other ancillary-to-the-main-text-of-the-program content (such as delimited comments).

// more context, quoted from the next diff: //

This diff implements the infrastructure necessary for the parser to handle extra error tokens in a systematic way. Specifically: prior to this diff, the FFP took an ad-hoc approach to handling extra error tokens. Sometimes, it would eat them. In other cases, it would construct an 'error' syntax node, but it was unclear which node in the larger syntax tree that error should be associated with. Now, this diff allows the parser context to passively do the work of managing error tokens. It allows them to be treated as trivia: if an unexpected token is encountered and deemed extra, the parser can convert it into a list of trivia and prepend it to the leading trivia of the next valid token.

Reviewed By: ericlippert

Differential Revision: D5465083

fbshipit-source-id: 3ebee9ea148b411f9ae1de804a0008d0e9dfce5b