Augmenting parser context to carry 'extra' error tokens
Summary:
This diff implements the infrastructure necessary for the parser to handle extra error tokens in a systematic way. Specifically: prior to this diff, the FFP took an ad-hoc approach to handling extra error tokens. Sometimes, it would eat them. In other cases, it would construct an 'error' syntax node, but it was unclear which node in the larger syntax tree that error should be associated with. Now, this diff allows the parser context to passively do the work of managing error tokens. It allows them to be treated as trivia: if an unexpected token is encountered and deemed extra, the parser can convert it into a list of trivia and prepend it to the leading trivia of the next valid token.

Implementing this scheme required refactoring the parser context--now the context is back to being a record again. A lot of code winds up being duplicated, but the resulting experience for the end user is very simple: see D5465181, the next diff in the stack, for a look at these methods in action.

Differential Revision: D5465141

fbshipit-source-id: cebb44acb8f2e3d803e690f6aaa5d2092c7fc5e5