Using trivia and augmented parser context to fix eating-extra-token bug
Summary:
This diff demonstrates the capabilities of the infrastructure I added in the last two diffs. Specifically, it uses that infrastructure to retain an error token in the syntax tree that had previously just been skipped.

Next step: phase out the few cases where Error SyntaxNodes are used in favor of this approach. (T20362204)

Differential Revision: D5465181

fbshipit-source-id: 0db84b4df67c6051cb3d823cc81f7f1c16c799cb