Pure refactoring: renaming expect_[token] to require_[token]
Summary: I changed the names of these methods to avoid conceptual confusion with what the parser context does when it "expect"s tokens. I think "require" more accuately reflects what these methods do: if the token's not there, they mark it missing, throw an error, and continue the parse from the unexpected token. (Meanwhile, when the parser context "expect"s a token, it doesn't rigidly enforce when that token should show up--it just considers it a possibility sometime in the future.)

Reviewed By: alexchow

Differential Revision: D5519582

fbshipit-source-id: 07ce7bfefa84417fbb32e1fdb9764ff54a39c8c2