parse explicitly annotated function calls in AST parser
Summary:
Modified the Call node of the parser AST to have an additional
hint list. Their use is intended to mirror how we explicitly annotate
new class instantiations.

The AST parser now parses function calls of the form

foo<int, string>();

where foo is a generic function with 2 tparams.

This is diff 1 of 5 planned diffs. The next 4 are:
1. parse annotated funcalls in the FFP.
2. modify the named ast to propagate type hints forward.
3. Add a NastCheck that checks the arity of the provided annotations matches
the arity of the function.
4. Add the typing, consuming the type hints.

The parsing has been split in two because this already touches a lot of files.

A unit test that verifies the parsing has been added.

Reviewed By: jamesjwu

Differential Revision: D5505575

fbshipit-source-id: 5f6659f23326c5a9fe2d96be2ec2d119d27edc40