generate different code for suspend calls in tail positions
Summary:
Tail position is defined as a final action performed in the function. In our case it will be:

- immediate expression of a return statement that is not enclosed inside the try-body of the try statement
- expression in parentesized expression located in a tail position
- (*) ? consequence or alternate parts of conditional expression located in a tail position
- (*) ? right hand side of null-coalescing operator located in a tail position
- (*) ? right hand side of the logical (||) and (&&) operators

when `suspend f()` appears in a tail position it can be emitted as `f($outer_continuation)` without creating addtional states in the state machine.

This diff adds support for the first two points, since currently lowering for conditional operator, null-coalescing operator and short-ciruiting operators is not yet implemented.

Reviewed By: michaeltingley

Differential Revision: D5502727

fbshipit-source-id: 78ea684ab390e92afa2efdcc2d253a1227810a39