Improve HHBBC inference of specialized array types on properties
Summary:
HHBBC currently doesn't attempt to track specialized array types on anything but
locals. There doesn't seem to be any real reason for this restriction so instead
expand the logic to handle to handle all possible base types. This actually
simplifies the logic a bit as there's less special cases.

Unfortunately for non-static properties, we must still drop any specialized
array type information when we're done. This is because an object may be
serialized, and we don't currently enforce that the array structure is kept
across the serialization. This isn't a problem for properties marked
AttrNoSerialize (like memoize caches), so we don't need to drop the specialized
information for those.

In some cases the inferred types aren't as good as they could be. This is
because we track property types in a flow-insensitive manner and our specialized
array types don't handle that well.

Reviewed By: markw65

Differential Revision: D5235014

fbshipit-source-id: 8adc0b6a45691d16a82993d061eb76c32158fa6b