Improving keyword autocomplete for type keywords
Summary:
- Added support for nullable types
- "void" and "noreturn" will only be completed in the return type annotation and not in class properties
- Type names are valid in exactly 4 places:
1. function/method returns (after a colon)
2. parameters (before the parameter name, after a comma or paren)
3. class properties (after a visibility modifier, const, or static)
4. Inside a generic type, i.e. array<int>

Void and noreturn are separate since they can only be return types and not property types. Void also has the special case where it can be used in Awaitable<void>, so we allow it inside any angle brackets.

"This" is only an allowed return type inside a class body.

Reviewed By: arxanas

Differential Revision: D5492274

fbshipit-source-id: f6e47cc6a9f996f272b27fc0fcf1cda3d20821d2