GC Manager for APC
Summary:
APCGCManager class to manage all to-be-freed APC * UNCOUNTED * values.
APCGCManager maintains a mapping : {allocation -> root}. root is the APCHandle* of the top-level APC object.
For example, if $a = ["str1",$b], $b = ["str2"]. We will have 4 entries in the treeMap:
1. { [StringData,ArrayData], $a }
2. { "str1", $a }
3. { [StringData], $a }
4. { "str2", $a }
For this reason, we won't do partially free on an APC object -- either don't free any allocation, or free the whole object.

APCGCManager maintains mapping: {rootAPCHandle -> visibleFlag} to implement GC for APC. When the usedAPCMemory exceeded a bar, it invokes a global GC scan and clean all invisible data.

For current version, the heap collector will invoke GC for APC when running hhvm in script mode. So it should be able to prevent OOM and pass the two unit tests when running in script mode(or in single thread). Run it with runtime flag -Eval.GCForAPC=true. Following things will happen:
1. Heap collector will call APCGCManager::sweep() during sweeping stage. This need to be removed after global GC API is provided.
2. Heap collector will call APCGCManager::mark() during marking stage.
2. Treadmill will call the traverse method to register {allocation, rootHandle} with APCGCManager when apc_delete()/apc_update() called.
3. Treadmill will ask APCGCManager::freeAPCHandles() to free data instead of freeing them by itself during treadmill round.
4. APCHandle::Create() will call APCGCManager::newAPCValueCreated() to maintain memory usage.

But nothing get changed in server mode, because there is no API to invoke global GC yet. We need that API to implement GC for APC in server mode.

Reviewed By: kodafb

Differential Revision: D5401852

fbshipit-source-id: baac13e56fc20836e53b9ab3bec25a31a149b4f7