Skip over unsafe tag for expression statements in codegen
Summary:
CodeGen did not have a definition for `Unsafeexpr` in expression statements. By itself, this isn't a problem, because the further call to `emit_expression` should codegen just fine. However, when we get a
```
/* UNSAFE_EXPR[2011]: This is safe */
await foo();
```
the codegen for expressions considers this a "complex await expression" and gens a NYI comment.

Reviewed By: vladima

Differential Revision: D5526608

fbshipit-source-id: 210e5a3d1bdf2b56dda664bdac932f631718c743