Improving error recovery on missing value in foreach statements
Summary:
This diff uses the parser context to fix another case in which deletion of a single token used to cause tons and tons of cascading errors. It also makes progress towards resolving the question "how do the expect_[token] methods inform the role of the parser context?"

**What I did:**

I noticed that deleting '$value' from the code in test_missing_foreach_value.php would cause the cascade of errors pasted at the end of this summary. Looking into it, I found that this happened because the expression parser thought that the RightParen was being used to generate a subscript_expression to follow the '=>' token, and the parser just kept digging itself deeper into a hole as a result. Now I've updated it so that a ')' is pushed onto the context immediately before the foreach_value is parsed. And, I've added a check-for-encountering-a-kind-that's-expected-later case to the match of parse_term in the expression parser. These two changes together were sufficient to nudge the error recovery here back on track :)

For more detail, see the before-and-after s-expressions below.

//Aside://

Notice that the RightParen is pushed onto the context as a member of a list called required_delimiters, rather than as a member of a list called recovery_tokens (like in D5501483). This is because I've noticed that tokens we want to include in the context come in two flavors: those that are merely allowed later, like 'class', and those that our program absolutely requires later in order to parse correctly, like ')' in some cases. I haven't figured out if I should do anything with this distinction, but for now I'll call the respective recovery lists 'recovery_tokens' and 'required_delimiters' just to keep track of how often the two types of cases occur.

//After://

                  ((foreach_statement 167
                  ((((whitespace 4)) foreach 7 ()) (() "(" 1 ())
                   (variable 9 ((() variable 9 ((whitespace 1)))))
                   (missing 0 ()) (() as 2 ((whitespace 1)))
                   (variable 4 ((() variable 4 ((whitespace 1)))))
                   (() => 2 ((whitespace 1))) (missing 0 ())
                   (() ")" 1 ((whitespace 1)))

//Before://
                  ((foreach_statement 161
                  ((((whitespace 4)) foreach 7 ()) (() "(" 1 ())
                   (variable 9 ((() variable 9 ((whitespace 1)))))
                   (missing 0 ()) (() as 2 ((whitespace 1)))
                   (variable 4 ((() variable 4 ((whitespace 1)))))
                   (() => 2 ((whitespace 1)))
                   (subscript_expression 131
                    ((() ")" 1 ((whitespace 1)))
                     (() { 1
                      ((whitespace 1) (single_line_comment 19)
                       (end_of_line 1)))
                     (binary_expression 101
                      ((subscript_expression 30
                        ((variable 16 ((((whitespace 6)) variable 16 ())))
                         (() [ 1 ())
                         (cast_expression 12
                          ((() "(" 1 ()) (() string 6 ()) (() ")" 1 ())
                           (variable 4 ((() variable 4 ())))))
                         (() ] 1 ((whitespace 1)))))
                       (() = 1 ((end_of_line 1)))
                       (function_call_expression 60
                        ((qualified_name 13
                          ((((whitespace 8)) qualified_name 13 ())))
                         (() "(" 1 ())
                         (list 45
                         ...

And it just kept going, generating an extremely erroneous parse tree all the way till the end of the file.

//The previous cascade of errors://

                  (6,36)-(6,36) An expression is expected here.
                  (8,69)-(8,69) A right brace is expected here.
                  (10,12)-(10,12) A class member, method, type, trait usage, trait require, xhp attribute, xhp use, or xhp category is expected here.
                  (10,16)-(10,16) A class member, method, type, trait usage, trait require, xhp attribute, xhp use, or xhp category is expected here.
                  (10,23)-(10,23) A class member, method, type, trait usage, trait require, xhp attribute, xhp use, or xhp category is expected here.
                  (10,24)-(11,1) A class member, method, type, trait usage, trait require, xhp attribute, xhp use, or xhp category is expected here.
                  (11,11)-(11,11) A class member, method, type, trait usage, trait require, xhp attribute, xhp use, or xhp category is expected here.
                  (11,14)-(11,14) A class member, method, type, trait usage, trait require, xhp attribute, xhp use, or xhp category is expected here.
                  (11,15)-(11,15) A class member, method, type, trait usage, trait require, xhp attribute, xhp use, or xhp category is expected here.
                  (11,31)-(11,31) A class member, method, type, trait usage, trait require, xhp attribute, xhp use, or xhp category is expected here.
                  (11,32)-(12,1) A class member, method, type, trait usage, trait require, xhp attribute, xhp use, or xhp category is expected here.
                  (12,6)-(12,6) A class member, method, type, trait usage, trait require, xhp attribute, xhp use, or xhp category is expected here.
                  (12,7)-(13,1) A class member, method, type, trait usage, trait require, xhp attribute, xhp use, or xhp category is expected here.
                  (15,10)-(15,10) A semicolon is expected here.
                  (15,17)-(15,17) A variable name is expected here.
                  (26,2)-(27,1) An expression is expected here.
                  (6,34)-(8,68) Use of the '{}' subscript operator is deprecated;  use '[]' instead.

Differential Revision: D5510120

fbshipit-source-id: 6ed11316a1e09e452045f4362578cdb3cc72318f