Fixing locations of parse-time errors
Summary:
In this diff, I modified the locations associated with certain errors so that they'd better hew to the design principle of "treat locations of errors as places you'd want to put a red underline in the code."

Meaning: I added a parameter to the with_error method to make it so that parse-time errors can now either precede a token, or encompass the token. (Previously they could only precede the token.) I've triggered this encompass-the-token behavior on error1055, error1056, and error1057 (reproduced below) but left locations reported on all other parse-time errors the same. (Those errors are the ones that start with a '1' in full_fidelity_syntax_error.ml, and they're all messages along the lines of "a ___ was expected here." I think it makes sense to report these as occurring on trivia in front of the text of the following token, because that following token is possibly not itself in error.)

The errors that are affected are:

  let error1055 = "A fallthrough directive can only appear at the end of" ^
    " a switch section."
  let error1056 = "This token is not valid as part of a function declaration."
  let error1057 text = "Encountered unexpected token '" ^ text ^ "'."

Differential Revision: D5531056

fbshipit-source-id: 24a12e5f4e7f01b614e9a2fd1fed935db75afc1f