changing lsp integration tests to always emit a .observed.log file.
Summary:
To make it easier to troubleshoot errors and update tests when breaks occur this change logs
what actually happened communicating with the LSP in all cases.  This should make it possible
to copy these over the expected files in cases where a change is made to the LSP that isn't
actually an issue, and the test needs to be updated.

Reviewed By: ljw1004

Differential Revision: D5555741

fbshipit-source-id: 70f2c6b5e52b58b36321a64ea9b7f3a6cbab9fdb