Streamline freeLocalsHelpers
Summary:
decRefHelper is structured like this:

   mov    (%rsi),%rdi
   cmpl   $0x1,(%rdi)
   jl     skip
   je     destroy
   decl   (%rdi)
   retq
  destroy:
   push   %rdx
   push   %rsi
   shr    $0x2,%ecx
   callq  *0x12e0e500(,%rcx,8)
   pop    %rsi
   pop    %rdx
  skip:
   retq

Because the conditional branches are forward branches, and we don't do
relocation for these stubs, they both end up being 6 byte branches,
which pushes the first retq out to offset 0x14. This means that even
for the fastest path, x64 will do two fetches.

By enabling relocation, we pull the retq in to offset 0x0c, and the
fast path only requires one fetch. But for uncounted entities, we
branch to the second retq, and still require two fetches. So move the
skip label to before the first retq, instead of the second.

This was followed by a loop:

   lea    -0x80(%rbp),%rdx
  loop:
   movzbl 0x8(%rsi),%ecx
   cmp    $0x39,%cl
   jle    .1
   callq  decRefHelper
  .1:
   add    $0x10,%rsi
   cmp    %rsi,%rdx
   jne    loop

again, the jle was 6 bytes, but can be 2 if we enable
relocation. Finally this was followed by 8 copies of

   movzbl 0x8(%rsi),%ecx
   cmp    $0x39,%cl
   jle    .1
   callq  decRefHelper
  .1:
   add    $0x10,%rsi

and again, relocation can fix the jle's

After these changes, the whole thing goes from just fitting in 4 cache
lines, to almost fitting in 3 - so I changed it to only unroll the
last 7 iterations of the loop, and got it down to 3.

But now we potentially touch an extra cache line, because decRefHelper
comes first, and the rest of the code always falls through to the very
end. By putting it last we can get away with 1 cache line for up to
two locals, 2 cache lines for up to 5, and 3 cachelines for the rest.

Reviewed By: mxw

Differential Revision: D5554479

fbshipit-source-id: f892a8a7e3f7d9daa5b68123af777683bf511461