Added a --stripped-tast option
Summary: This adds a --stripped-tast which first generates the typed AST, then strips the type information so that it has the structure of a regular named AST, then prints out that named AST. This is used for finding holes in the typed AST where expressions were dropped. Eventually it can be used for automatic unit tests by ensuring that --nast and --stripped-tast always print out the same value for a given source file.

Reviewed By: jamesjwu

Differential Revision: D5521937

fbshipit-source-id: d535f9d9a3748db22ebbd1b80a087c0f1757de36