Always use size class vs requested bytes
Summary:
When accumulating memory usage, always use the actual size class
of the memory block, rather than sometimes using the requested
size (depending on the allocation api), so that internal fragmentation
is included in "usage" metrics, and the OOM check.

This is already the way jemalloc works, and how we deal with "big"
allocations; this diff makes it work consistently for small objects
as well.

Reviewed By: alexeyt

Differential Revision: D5457249

fbshipit-source-id: 37652c1942ef8d9b6489321112ccc265f1f168a5