Reduce hhbbc lock contention
Summary:
perf reported that we were spending 36% of the time in the kernel
waiting on locks, and this one seemed to be the worst offender.

I think the big issue was that create_func_info was written as if
creation was the normal case, but in fact its mostly used for
lookup. This means we would create and destroy a FuncInfoValue on
every lookup, while holding the lock. Just fixing that would probably
have helped a lot, but using concurrent_unordered_map as well avoids
the need for the lock at all.

Reviewed By: ricklavoie

Differential Revision: D5564726

fbshipit-source-id: 2c5cbb7b0dfac7f2881943fcbb1e022a20b21273