Add a note about false positive on need for server restart
Summary:
The monitor is supposed to be always responsive. So, code in clientConnect and clientStart which check for the monitor's responsiveness doesn't bother with retries or timeouts.

But there's a race during startup. Imagine if you have an automated background task which monitors hh_server and restarts it if it goes down. And imagine that at the keyboard you do "hh_client restart". Once of the two processes will be the first to start it up, and there'll be a period of 1-2 seconds during which it replies ECONNREFUSED to the other process.

Currently the other process takes this as a sign that the monitor is in an unhealthy state and should be restarted. So it will shut it down nicely (which it does via a timeout) and then restarts it\! When all it should have done is wait.

I haven't implemented retry+timeout. That's because the code is too tangled up at the moment. We need a long retry to wait until the post-handoff work has done (i.e. for loadstate or typechecking to finish), but only a short ~3s retry to wait until the connection to the monitor has been established. The ClientConnect.connect method currently does both those things and so doesn't have a good place to distinguish the retries.

So I'll make do with merely documenting the scenario for now.

Reviewed By: alexchow

Differential Revision: D5132238

fbshipit-source-id: 792c18314778a28f07458a0bf3b4badc590ff65d