Use a vector of FuncInfos
Summary:
Even after D5564726 we were spending quite a bit of time looking up
FuncInfos. There's always (eventually) a 1-1 correspondence between
php::Funcs and FuncInfos, so just add an index to php::Func (it fits
into an existing padding hole without increasing the size of the
struct), and use a vector of FuncInfos.

Reviewed By: ricklavoie

Differential Revision: D5571274

fbshipit-source-id: 2905b68e2eda470d6e2f1477c12fa54884b55e36