emit 'busy:initializing' over LSP
Summary:
This introduces a Nuclide-specific extension to LSP, to let the LSP server tell the client to display a busy-indicator, or to display an action-required indicator.

If the client says it has the capability to display such an indicator, then hh_client will use it to indicate "busy" for the time after initialization has returned until the hh_server is ready.

Reviewed By: dabek

Differential Revision: D5524246

fbshipit-source-id: b4c3a8a7fd4a9a9e29878da56135d9ea9ef6fd40