Open/close shouldn't trigger a lazy recheck
Summary:
Currently when you open/close a file, hack triggers a lazy recheck.

This diff makes it so that if you open/close a file, and the contents on disk are the same as the contents from the editor buffer, then it skips the lazy recheck.

This opens the door to a user-relevant instruction that in all remaining cases when a lazy recheck has finished, then they should also trigger a global recheck to see program-wide errors.

Note: there was an additional oddity in the previous code. Imagine you have a file with unsaved changes and you close it. Previously the code said "let's put the unsaved editor contents into the FileHeap cache". You'd think this would be a bug\! But it turns out not, because the previous code also put that file into ide_to_parse; later on, in serverTypeCheck.ml:parsing, the file showed up in the variable "disk_files". And it removed all disk files from the FileHeap. So the FileHeap cache for this file ended up being ignored.

Reviewed By: dabek

Differential Revision: D5541532

fbshipit-source-id: 2835bec90ffe5e2d49ee8d8d7540c930e0a3a7dd