Added Efun nodes to the typed AST (except variadic Efuns)
Summary:
Right now, Efun nodes are missing from the TAST. This adds them back. Variadic efuns don't work yet, because we actually never typechecked those args in the first place. They will be fixed in a diff on top of this one.

For example, this function typechecks, but should error.

```
<?hh //strict

function shouldError(): int {
  $op = function(int ...$i) {
    return $i[0] + 3;
  };

  return $op("A string.");
}
```

Reviewed By: dlreeves

Differential Revision: D5503210

fbshipit-source-id: b78e254c1309a9026e3db7f1392a88330612a3c9