Simplify errors with generic expression dependent types
Summary:
This diff does two things:

First, it refactors the optional arguments to unify into a "unify_options" data structure. To preserve the unify_ref behavior we have with Typing_utils to avoid cyclic dependencies, we define this data structure in Typing_utils and thread it through.

Second, it adds a new unification option called "simplify_errors". Simplify_uerror calls unify, which in turn may call simplify uerror. This behavior is made worse by the fact that we don't properly thread through most of the unification options. To make the approach more sensible, we pass this flag in when we want unification to try to resimplify the errors by calling itself. This way, we only ever simplify unification errors once.

Lastly, this diff enables simplifying errors for generic expression dependent types.

Reviewed By: dlreeves

Differential Revision: D5435470

fbshipit-source-id: 55003e447fa620c42726e68a1a91f06bd617c737