go to constructor, go directly to constructor, do not pass class, do not collect $200
Summary:
This diff makes it quicker and easier to command-click on a `new Cloo()` call and be taken directly to its constructor.

BACKGROUND:

Prior to LSP, if you had `new Cloo()` and did hyperclick on the word `Cloo`, it used to offer you a popup with two options:

  \Cloo (Directory\Subdirectory\Cloo.php)
  \Cloo::__construct (Directory\Subdirectory\Cloo.php)

Curiously it did this even if `Cloo` itself lacks a constructor. It still showed an entry called `\Cloo:__construct`, but clicking on it took you to the constructor of the base class.

With the switch to LSP, LSP no longer provides names human-readable names `\Cloo` or `\Cloo::__construct`. hansonw has made it so the Atom popup will just display filenames+lines in the popup:

  Directory\Subdirectory\Cloo.php:5
  Directory\Subdirectory\Cloo.php:7

VSCode doesn't use a popup. Instead it displays an inline peek window where the two options are visible and you can chose between the two of them.

SOLUTION:

In fact, neither the VSCode nor the Atom solution are all that great. They all involve extra clicks. It'd be much slicker to just go directly to where you want to go.

I'll wager that "where you want to go" is usually the constructor. But that even if it's not, then going direct to the constructor is better than being asked to chose.

Related: in Typescript, when you command-click on a call to the constructor, it takes you direct to the constructor. (In the case where `Cloo` lacks a constructor but its base class does, then it takes you direct to that base class constructor, so you can see what the parameters are). Even if you really wanted to go to the class declaration, well, the constructor is so close to the class declaration that you'll probably see it anyway.

Related: in Flow, when you command-click on a call to the constructor, it takes you to the class declaration rather than to the constructor.

This diff copies the Typescript behavior: if a go-to-definition would return both "class" definition location (for the class) and "method" definition location (for the constructor), it gives only the method location.

Reviewed By: arxanas

Differential Revision: D5569199

fbshipit-source-id: ab2ec126d98d7e1b83be941324b262499436a0e9