fixing alias/trait parsing stuff
Summary:
Make the hhas parser work for a richer subset of use directives.

Note: I think that the visibility modifiers in class_use_aliases should probably be
a list/set, rather than a single option. It'd be nice if someone fixed that in a
subsequent diff...

This fixes some iffy code in D5446972.

Reviewed By: oulgen

Differential Revision: D5574129

fbshipit-source-id: 1f89595a5ee77aa8c42137b3bef82e2a9b2ab43e