Type soundness with contravariance and unresolved
Summary:
We often generate fresh type variables and initially bind these to an empty union `Tunresolved []`. Attempting to subtype against a concrete supertype (which often happens through contravariance) fails to produce an error. For example, consider the following code:
```
interface I1<-T> { public function f(T $t): int; }
final class C1 implements I1<string> {
  public function f(string $t): int { return 42; }
}
final class C2<T> {
  public function __construct(private I1<T> $impl) {}
  public function f(T $t): int { return $this->impl->f($t); }
}
function BreakIt(): void {
  $c1 = new C1();
  $c2 = new C2($c1);
  // the following line fails in runtime
  $c2->f(42);
}
```
Here is what happens inside the type checker. `$c2` is assigned type `C2<v>` for fresh variable `v` initially bound to `Tunresolved []`.
Now we check `C1 <: I1<v>` and so `v <: string`. But the latter is just accepted because `v` is expanded to `Tunresolved []`.

The long term solution is to move from unification-based type inference with the "growing" of unions, to a system based on subtype constraints. This diff is a first step. When asked to subtype `v <: t` for some variable `v` that is bound to `Tunresolved []`, we add a deferred check `v <: t` to be run later, from the existing "todo" list that is used mainly for generic parameter constraints. In the above example, `v` will get unified to `int` and so when the subtype is checked later, because `int` is not compatible with `string`.

For now, this is under an experimental feature flag. It's switched on by default for `hh_single_type_check`, but for `hh_client` you need to add
```
enable_experimental_tc_features = unresolved_fix
```
to the `.hhconfig` file, along with any other experimental flags.

Reviewed By: dlreeves

Differential Revision: D4380897

fbshipit-source-id: 23d7bd11a29cd6e2ff5d2ca51dc0f57433e25438