Fix potential use after frees when storing keysets in APC
Summary:
SetArray::MakeUncounted had some confusion where it was turning the
string keys of the *input* keyset into uncounted values, instead of the new
uncounted keyset. As a result, we'd create an uncounted keyset which could
contain counted strings. This keyset could sit in APC across requests, long
after the strings it contains have been reclaimed. When we eventually retrieve
the keyset out of APC and access the strings, we end up accessing freed memory
(or perhaps memory reclaimed for other values).

Give the variables in SetArray::MakeUncounted sane names to avoid confusion over
which is the input and which is the output. Add a test case which triggered this
use-after-free. To make the test fail more deterministically, add a new runtime
option to trash fill the request heap as the request is exiting. This ensures
that any references to request allocated values will see trashed memory and
trigger assertions more quickly.

Reviewed By: mxw

Differential Revision: D5588897

fbshipit-source-id: 3dfb239aaa1fbdce5c25118f78ce1bca6bf1598d