fix crash when ColFromArray doesn't have an array on the stack
Summary: HHVM was crashing when the argument on top of the stack to ColFromArray was not a hack array (Vec or Dict). Additionally it was crashing when trying to create an (Imm)Vector from a Dict or an (Imm)Set or an (Imm)Map from a Vec. This adds runtime checks that this does not occur, since it can't be checked in the verifier.

Reviewed By: alexeyt

Differential Revision: D5578542

fbshipit-source-id: b5de4619c84c88f520bc852dcca638be2a36560a