Improving error recovery on incomplete namespace declarations
Summary:
In this diff, I improve the error recovery on improperly declared namespaces. I originally started this fix because I'd noticed the TODO and thought it could be solved with the parser's context, but after giving it more thought I realized that...since namespaces are only legal at declaration level anyway, and they contain only declarations, directly recovering back to declaration level is a reasonable error recovery strategy, since in either case a namespace must be immediately followed by more declarations.

I also modified the text of the relevant error--SyntaxError.error1038--to be more representative of the updated recovery strategy.

Strategy: I chose to make the parser recover by throwing an error, and then pretending that an improperly declared namespace is just

    namespace NS { }

but with all the braces missing. I could have instead chosen to make the parser pretend the namespace was

    namespace NS;

but with the semicolon missing. I figured the first option was better because it "impacts less of the file" by just pretending that the erroneous line isn't there at all, but I think there are compelling reasons to go the other way too. If you have any feedback on this decision, let me know\!

Regarding #impact: previously, namespace errors weren't terribly cascading, but they also weren't very targeted or informative, and the locations were all wrong. Below are the before-and-after error messages for the test I added in this diff--note that the "after" do a much better job of pointing out the root cause of the errors.

//Before//

    (4,9)-(4,9) A namespace body is expected here.
    (4,15)-(5,1) A semicolon is expected here.
    (28,9)-(28,9) A namespace body is expected here.
    (28,14)-(29,1) A semicolon is expected here.
    (31,35)-(32,1) An expression is expected here.
    (31,35)-(32,1) A semicolon is expected here.
    (36,50)-(37,1) A right brace is expected here.

//After//

    (3,61)-(4,1) A semicolon or a namespace body is expected here.
    (27,61)-(28,1) A semicolon or a namespace body is expected here.
    (31,35)-(32,1) An expression is expected here.
    (37,1)-(37,1) A right brace is expected here.

Differential Revision: D5558701

fbshipit-source-id: 40b8432c99dd16b3105130544f8f2445c9321a0c